requirements.txt
# Backend зависимости
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
requests>=2.31.0
python-dateutil>=2.8.0

# Frontend зависимости (PyQt6) - требует Visual Studio Build Tools
# PyQt6>=6.6.0
# PyQt6-sip>=13.0.0
# PyQt6-Qt6>=6.6.0

# Дополнительные зависимости
wheels>=0.42.0
setuptools>=65.0.0

# Excel/CSV экспорт/импорт
openpyxl>=3.1.0
